PAR-2522 (#1197)
* PAR-2522 Test GitHub 'migration' mini release
Add full stop to the description for Par Data Test.
Add the two development environment configuration directories for Docksal and DDEV.

* PAR-2522 Andrew Crook Test
1. Update the database namespace configuration parameter and add the 'autoload' which is added on installation these days.
2. Correct the spelling mistake.
3. Use the value of scheme passed through VCAP_SERVICES if present.
   By default the redis Docker images endpoints do not have TLS functionality.
4. When running command line drush, need to reformat the VCAP_SERVICES environmental variable when running under Docksal.

* PAR-2522 Andrew Crook Test
Export the view so that it does not appear in the synchronize list as difference.

* PAR-2522 Test GitHub 'migration' mini release
Add another full stop to the description for Par Data Test.

* PAR-2522 Test GitHub 'migration' mini release
See if the Smoke test failure is due to parallel running.

* PAR-2522 Test GitHub 'migration' mini release
revert to parallel running.

---------

Co-authored-by: Andrew Crook (Somersoft) <andrew.crook@somersoft.co.uk>