PAR-2543 Site addons services check Base upon the release March 25 branch (#1220)
* JIRA PAR-2543
Update the code so that more of the test pass.
Conform to Code Styling guide.

* JIRA PAR-2543
Update the test code so that more pass as the test were behind the changes to the Entities definitions.
Update phpunit configuration file.

* JIRA PAR-2543
Update the test code so that the correct value is passed to the assert.

* JIRA PAR-2543
As this file was involved in the checking of tests
it was updated and now passes PHPCS checks.
There has been no change in functionality.

* JIRA PAR-2543
As this file was involved in the checking of tests
it was updated and now passes PHPCS checks.
There has been no change in functionality.

* JIRA PAR-2543
Update the setup so to check that the created user does not have uid of 1
 as this still is checked when permissions are checked.
See SuperUserAccessPolicy::calculatePermissions()

* JIRA PAR-2543
Update the tests so that they check the fields that have constraints added to them.

* Correct which files are to be ignored.