JIRA PAR-2588/-2597/-2598 Address Zap report findings Medium and Low (#1247)
* JIRA PAR-2588 Address Medium (125) CSP: Failure to Define Directive with No Fallback
Move the header configuration from HTTP Response Headers (http_response_headers) to Security Kit (seckit) as it has a better UI.
Extended the CSP configuration to include other Directives than 'frame-ancestors'
It is still in report only mode to pick up if there are other Directives need to be updated as only the pages used by the Cucumber testing have been visited.

* JIRA PAR-2598 Address Low  (131) Server Leaks Information via "X-Powered-By" HTTP Response Header Field(s)
Remove the header.

* JIRA PAR-PAR-2597 Address Low (10) Cookie without SameSite Attribute
Updated the services.yml file by comparing it with the default.services.yml and taking across the missing configuration.
This includes the same site cookie option.
The security policy around user 1 is not default value of true but false, so they their permissions is the same as other accounts.

* JIRA PAR-PAR-2597 Address Low (10) Cookie without SameSite Attribute
Update the non production as well.

* JIRA PAR-2598 Address Low (131) Server Leaks Information via "X-Powered-By" HTTP Response Header Field(s)
Remove the header.