PAR-2603 Profile Invite not being sent (#1264)
* Latest working changes for running the Cucumber test locally under Docksal.

* JIRA PAR-2603 profile invite not being sent.
Add the missing configuration file so for the test maillog is the default mail handler.

* JIRA PAR-2603 profile invite not being sent.
Check to see if there is a email address for the user.
User 0, the anonymous user does not have one and this was causing problems in the SQL code as NULL and not a character sequence was being passed.

* Updates after merging latest master.

* Correct the documentation.

* JIRA: PAR-2603 Profile Invite not being sent
Update the access check functionality so that it returns the correct value and not always allowed.
Only print out the link header information if there is a link.
Both have changes to remove PHPCS warning and errors.

* Increase the numbers of processes.

* Delete the updated membership files as well.
Rebuild the search index so that it is fully populated.

* Always run the regression tests.

* Create a system variable call app.dir and point that at /app where the volume is mounted and use that combination to get the files.

* remove the .docksal directory