PAR-2600 Anti Virus (#1265)
* JIRA PAR-2600 Lack of anti virus on V3
Restore the functionality using changes from
https://github.com/OfficeForProductSafetyAndStandards/beis-primary-authority-register/commit/56dc14deaed3766463564333720fde7ed6363a74

* JIRA PAR-2600 Lack of anti virus on V3
Delete old patch file.

* JIRA PAR-2600 Lack of anti virus on V3
Add setting of the environment variables.

* PAR-2600: Updating ClamAV settings for prod and non-prod

* PAR-2600: Removing devel and ctools from prod

* NO TICKET: Test fixes

* PAR-2600: Deployment config fixes

* PAR-2600: Config sync with prod

* JIRA PAR-2600 Lack of anti virus on V3
Sort the classes out so that it is correct.

* PAR-2600: Updating build task to set env vars post deploy

* PAR-2600: Typo fix

* JIRA PAR-2600 AV
Update the changes for the dev and test environments to use the staging clamAV services.

* JIRA PAR-2600 AV
Update the ClamAV patch.
Update the configuration splits.

* JIRA PAR-2600 AV
Update the configuration splits.

* PAR-2600: Composer update

* JIRA PAR-2600 AV
Update the patch so that the endpoint will use the correct form input field.e

* JIRA PAR-2600 AV
Update the order so that it does not appear on Configuration synchronize page.

* JIRA PAR-2600 AV
Corrected the index names.

* JIRA PAR-2600 AV
Add masquerade.settings.yml for dev.

* JIRA PAR-2600 AV
Rerolled the patch as lines had been shortened.

---------

Co-authored-by: Robert Wild <robert.wild@affinity-digital.com>
Co-authored-by: Robert Wild <162973285+robwild-affinitydigital@users.noreply.github.com>